Added throw statement to database, rewrote 'BotCommands()'

+++ program/index.js
@@ -30,18 +30,34 @@ client.on("ready", () => {
 
         switch (command = args[0]){
             case '!help':
-                commandResponse = GetBotCommands(args);
+                if (!Validator.validateArguments([args[1]])){
+                    commandResponse = GetBotCommands();
+                } else {
+                    commandResponse = GetBotCommand(args[1]);
+                }
                 message.reply(commandResponse.message ? commandResponse.message : commandResponse.error);
                 break;
             case '!addchannel':
+                if (!Validator.validateArguments([args[1], message.channel.id])){
+                    return message.reply("Invalid arguments in command.");
+                }
                 Database.AddChannel(args[1], message.channel.id)
-                    .then(response => { message.reply(response); })
-                    .catch(error => { message.reply(error); });
+                    .then(() => { message.reply("Channel successfully added!"); })
+                    .catch(error => {
+                        console.error(error);
+                        message.reply("Something went wrong while adding channel.");
+                    });
                 break;
             case '!removechannel':
+                if (!Validator.validateArguments([args[1], message.channel.id])){
+                    return message.reply("Invalid arguments in command.");
+                }
                 Database.RemoveChannel(args[1], message.channel.id)
-                    .then(response => { message.reply(response); })
-                    .catch(error => { message.reply(error); });
+                    .then(() => { message.reply("Channel successfully removed."); })
+                    .catch((error) => {
+                        console.error(error);
+                        message.reply("Something went wrong while removing channel.");
+                    });
                 break;
             case '!getupdate':
                 if (!Validator.validateArguments([args[1]])){
@@ -50,7 +66,7 @@ client.on("ready", () => {
                 message.reply("Getting update article, please wait...");
                 Scheduler.SendUpdate(args[1],[message.channel.id])
                     .catch((error) => {
-                        console.log(error);
+                        console.error(error);
                         message.reply("Something went wrong while getting update.");
                     });
                 break;
@@ -81,41 +97,35 @@ function SendMessage(channel, message, embed = false){
     else { client.channels.get(channel).send(message); }
 }
 
-function GetBotCommands(args) {
-    let output = {}, messageBody = "";
-    let helpAboutCommand = args[1];
+function GetBotCommands() {
+    let output = "";
 
-    /* If second parameter is not given */
-    if (!helpAboutCommand) {
-        messageBody = "To learn more about a command, type \"!help <command>\".\n";
-        messageBody += "Here are all of my commands:\n\n";
-        for (let key in constants.BotCommands) {
-            messageBody += constants.BotCommands[key]["command"] + "\n";
-        }
-    } else { /* Second parameter given */
-        let command;
-
-        /* Remove exclamation mark */
-        if (helpAboutCommand.indexOf("!") === 0) {
-            command = helpAboutCommand.substr(1)
-        } else {
-            command = helpAboutCommand;
-        }
-
-        /* If command exists.. */
-        if (constants.BotCommands[command]) {
-            messageBody = "\nCommand:      " + constants.BotCommands[command]["command"] + "\n";
-            messageBody += "Description:    " + constants.BotCommands[command]["description"] + "\n\n";
-            messageBody += "Syntax:             " + constants.BotCommands[command]["syntax"] + "\n";
-            messageBody += "Example:          " + constants.BotCommands[command]["example"] + "\n";
-            messageBody += "Possible argument values:\n" + constants.BotCommands[command]["argvalues"] + "\n\n";
-            messageBody += "If a command parameter includes a question mark (?), that parameter is optional.";
-        } else {
-            return new InvalidArguementsError();
-        }
+    output += "To learn more about a command, type \"!help <command>\".\n";
+    output += "Here are all of my commands:\n\n";
+    for (let key in constants.BotCommands) {
+        output += constants.BotCommands[key]["command"] + "\n";
     }
 
-    output.message = messageBody;
+    return output;
+}
+function GetBotCommand(command){
+    let output = "";
+
+    /* Remove exclamation mark */
+    if (command.indexOf("!") === 0) { command = command.substr(1); }
+
+    /* If command exists.. */
+    if (constants.BotCommands[command]) {
+        output += "\nCommand:      " + constants.BotCommands[command]["command"] + "\n";
+        output += "Description:    " + constants.BotCommands[command]["description"] + "\n\n";
+        output += "Syntax:             " + constants.BotCommands[command]["syntax"] + "\n";
+        output += "Example:          " + constants.BotCommands[command]["example"] + "\n";
+        output += "Possible argument values:\n" + constants.BotCommands[command]["argvalues"] + "\n\n";
+        output += "If a command parameter includes a question mark (?), that parameter is optional.";
+    } else {
+        throw new InvalidArguementsError();
+    }
+
     return output;
 }
